Fix detection of deps dir for cover.spec files
It seems that collect_glob/3 does not correctly determine that
it should skip a cover.spec file inside a dependency directory
and it ends up seeing all available cover.spec files (the one
for the root project and the ones inside their deps). The fix
in this case is to try to detect cover.spec files that are
inside the deps directory by trying to match the start of the
absolute path against the deps directory in a simple way.